+++ changelog.rst
@@ -1,3 +1,25 @@
+version 0.5.0 2012-07-26 released
+changelog: (2012-07-26)
+In pracmatical usage, a key "message" for all responese is very convenient for development.
+So add 'message' key to "Optional Keys" for type(success, fail)
+
+
+| ''Type'' | ''Description''                                        | ''Required Keys''  | ''Optional Keys'' |
+|----------+--------------------------------------------------------+--------------------+-------------------|
+| success  | All went well, and (usually) some data was returned.   | status, code, data | message           |
+|          |                                                        |                    |                   |
+| fail     | There was a problem with the data submitted,           | status, code, data | message           |
+|          | or some pre-condition of the API call wasn't satisfied |                    |                   |
+|          |                                                        |                    |                   |
+| error    | An error occurred in processing the request,           | status, message    | code, data        |
+|          | i.e. an exception was thrown                           |                    |                   |
+
+
+
+
+
+
+
 version 0.4.0 2012-07-16 released
     fix: in pracmatical usage, code representing status is very
     import, so add 'code' key to "Required Keys" for "type" of